requirements.txt
annotated-types==0.7.0
anthropic==0.29.0
anyio==4.9.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.7
filelock==3.18.0
fsspec==2025.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.53.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.33.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.3.0
openai==1.26.0
orjson==3.10.18
packaging==25.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
slack-sdk==3.20.2
sniffio==1.3.1
starlette==0.37.2
tenacity==8.2.2
tokenizers==0.21.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.21.0
watchfiles==1.1.0
websockets==15.0.1
